+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## 0.0.1
-
-* TODO: Describe initial release.
+## 2.0.0
+- Break change, remove scrollable_positioned_list, add flutter_list_view
+- Refresh control will construct by user, like pull_to_refresh